infra: Add structure for GTS deinit/destroy
Even though it is not needed right now, the GTS now has a method to destroy/deinitialize itself.
This is only called when the GTS was initialized as an intermediate source to calibrate another higher accuracy timer.
Support for both timers and GTS to calibrate is removed for simplicity. Only a global time source can be used to calibrate timers.